planned meal services refactor

+++ lib/data/planned_meal/source/remote/firebase_planned_meal_service.dart
@@ -1,5 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:mealapp/common/helper/handle_firestore_operation/exception/exception.dart';
+import 'package:mealapp/common/helper/handle_firestore_operation/exception/handle_firestore_exception.dart';
 import 'package:mealapp/data/meal/mapper/meal_mapper.dart';
 import 'package:mealapp/domain/planned_meal/entity/planned_meal_entity.dart';
 
@@ -12,37 +14,48 @@ abstract class FirebasePlannedMealService {
 class FirebasePlannedMealServiceImpl implements FirebasePlannedMealService {
   @override
   Future<void> addPlannedMeal(PlannedMealEntity plannedMeal) async {
-    final user = FirebaseAuth.instance.currentUser;
-    await FirebaseFirestore.instance
-        .collection('Users')
-        .doc(user!.uid)
-        .collection('PlannedMeals')
-        .doc('${plannedMeal.date}_${plannedMeal.meal.mealId}')
-        .set({
-          'date': plannedMeal.date,
-          'meal': MealMapper.toModel(plannedMeal.meal).toMap(),
-        });
+    return handleFirestoreException(() async {
+      final user = FirebaseAuth.instance.currentUser;
+      if (user == null) throw UnauthorizedException();
+      await FirebaseFirestore.instance
+          .collection('Users')
+          .doc(user.uid)
+          .collection('PlannedMeals')
+          .doc('${plannedMeal.date}_${plannedMeal.meal.mealId}')
+          .set({
+        'date': plannedMeal.date,
+        'meal': MealMapper.toModel(plannedMeal.meal).toMap(),
+      }).timeout(const Duration(seconds: 15));
+    });
   }
 
   @override
   Future<void> removePlannedMeal(DateTime date, String mealId) async {
-    final user = FirebaseAuth.instance.currentUser;
-    await FirebaseFirestore.instance
-        .collection('Users')
-        .doc(user!.uid)
-        .collection('PlannedMeals')
-        .doc('${date}_$mealId')
-        .delete();
+    return handleFirestoreException(() async {
+      final user = FirebaseAuth.instance.currentUser;
+      if (user == null) throw UnauthorizedException();
+      await FirebaseFirestore.instance
+          .collection('Users')
+          .doc(user.uid)
+          .collection('PlannedMeals')
+          .doc('${date}_$mealId')
+          .delete()
+          .timeout(const Duration(seconds: 15));
+    });
   }
 
   @override
   Future<List<Map<String, dynamic>>> getPlannedMeals() async {
-    final user = FirebaseAuth.instance.currentUser;
-    final snapshot = await FirebaseFirestore.instance
-        .collection('Users')
-        .doc(user!.uid)
-        .collection('PlannedMeals')
-        .get();
-    return snapshot.docs.map((doc) => doc.data()).toList();
+    return handleFirestoreException(() async {
+      final user = FirebaseAuth.instance.currentUser;
+      if (user == null) throw UnauthorizedException();
+      final returnedData = await FirebaseFirestore.instance
+          .collection('Users')
+          .doc(user.uid)
+          .collection('PlannedMeals')
+          .get()
+          .timeout(const Duration(seconds: 15));
+      return returnedData.docs.map((e) => e.data()).toList();
+    });
   }
 }
